requirements.txt
python-telegram-bot
SQLAlchemy
python-dotenv
requests
celery
redis
